pyproject.toml
[tool.poetry]
name = "aws-merlin-agent"
version = "0.1.0"
description = "MERLIN: Marketplace Earnings & Revenue Learning Intelligence Network"
authors = ["MERLIN Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "aws_merlin_agent", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.130.0"
constructs = "^10.3.0"
streamlit = "^1.32.0"
pydantic = "^2.6.0"
requests = "^2.31.0"
boto3 = "^1.34.0"
botocore = "^1.34.0"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
scikit-learn = "^1.4.0"
xgboost = "^2.0.0"
aws-lambda-powertools = "^2.31.0"
pyspark = {version = "^3.5.0", optional = true}

[tool.poetry.extras]
glue = ["pyspark"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.3.0"
black = "^24.3.0"
moto = "^4.2.0"
localstack = "^3.0.0"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
